+++ CHANGELOG.md
@@ -5,6 +5,15 @@ project adheres to [Semantic Versioning](http://semver.org).
 
 ## Unreleased
 
+## 0.3.0 - 2018-04-07
+
+### Added
+
+  - Package configuration file.
+  - Public folder with template and favicon.
+  - Application index and service.
+  - Application, day and slot components.
+
 ## 0.2.0 - 2018-04-06
 
 ### Added